Drop versioned dependency on debianutils, satisfied since Bookworm
Bookworm shipped with debianutils 5.7-0.5, so a versioned dependency is no
longer required in Forky, which will release with Bookworm being oldoldstable.
Also debianutils is prio: essential, so no dependency is required at all.

This is also useful when minimizing read-only container images, so more
pkgs can be removed when building such images as the final step.